# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vera Nexus BI Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EventCategory = typeof EventCategory[keyof typeof EventCategory];


export const EventCategory = {
  technology: 'technology',
  finance: 'finance',
  markets: 'markets',
  health: 'health',
} as const;

export type EventSentiment = typeof EventSentiment[keyof typeof EventSentiment];


export const EventSentiment = {
  positive: 'positive',
  negative: 'negative',
  mixed: 'mixed',
} as const;

export interface Event {
  id: number;
  year: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: EventCategory;
  sentiment: EventSentiment;
  /** Impact score 0-100 */
  impact: number;
  source: string;
  rippleCount?: number;
  tags?: string[];
}

export interface RippleResult {
  eventId: number;
  analysis: string;
  causalChain?: string[];
  affectedSectors?: string[];
}

export type ReportCategory = typeof ReportCategory[keyof typeof ReportCategory];


export const ReportCategory = {
  technology: 'technology',
  finance: 'finance',
  markets: 'markets',
  health: 'health',
} as const;

export interface Report {
  id: number;
  title: string;
  source: string;
  publishedAt: string;
  category: ReportCategory;
  summary: string;
  /** @nullable */
  imageUrl?: string | null;
  tags?: string[];
}

export interface SummaryResult {
  reportId: number;
  summary: string;
  keyTakeaways: string[];
  causalImplications?: string;
}

export type CompanyCategory = typeof CompanyCategory[keyof typeof CompanyCategory];


export const CompanyCategory = {
  technology: 'technology',
  finance: 'finance',
  markets: 'markets',
  health: 'health',
} as const;

export interface Company {
  id: number;
  name: string;
  yearRange: string;
  description: string;
  category: CompanyCategory;
  tags?: string[];
  /** @nullable */
  failureReason?: string | null;
}

export interface AutopsyResult {
  companyId: number;
  rootCause: string;
  timeline: string;
  lessonsLearned: string[];
  causalChain?: string[];
  /** @nullable */
  analogy?: string | null;
}

export type SignalSentiment = typeof SignalSentiment[keyof typeof SignalSentiment];


export const SignalSentiment = {
  positive: 'positive',
  negative: 'negative',
  warning: 'warning',
  neutral: 'neutral',
} as const;

export interface Signal {
  id: number;
  name: string;
  /** @nullable */
  value: string | null;
  change: string;
  sentiment: SignalSentiment;
}

export type ThoughtCategory = typeof ThoughtCategory[keyof typeof ThoughtCategory];


export const ThoughtCategory = {
  technology: 'technology',
  finance: 'finance',
  markets: 'markets',
  health: 'health',
} as const;

export type ThoughtReactions = {[key: string]: number};

export interface Thought {
  id: number;
  content: string;
  author: string;
  category: ThoughtCategory;
  createdAt: string;
  reactions?: ThoughtReactions;
  tags?: string[];
}

export type ThoughtInputCategory = typeof ThoughtInputCategory[keyof typeof ThoughtInputCategory];


export const ThoughtInputCategory = {
  technology: 'technology',
  finance: 'finance',
  markets: 'markets',
  health: 'health',
} as const;

export interface ThoughtInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
  /** @minLength 1 */
  author: string;
  category: ThoughtInputCategory;
  tags?: string[];
}

export type ReactionInputReactionType = typeof ReactionInputReactionType[keyof typeof ReactionInputReactionType];


export const ReactionInputReactionType = {
  fire: 'fire',
  brain: 'brain',
  chart: 'chart',
  signal: 'signal',
  agree: 'agree',
} as const;

export interface ReactionInput {
  thoughtId: number;
  reactionType: ReactionInputReactionType;
}

export type ReactionResultAction = typeof ReactionResultAction[keyof typeof ReactionResultAction];


export const ReactionResultAction = {
  added: 'added',
  deleted: 'deleted',
} as const;

export interface ReactionResult {
  action: ReactionResultAction;
  thoughtId: number;
  reactionType: string;
}

export type AnalyzeInputSessionHistoryItem = {
  role?: string;
  content?: string;
};

export interface AnalyzeInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  businessContext?: string | null;
  sessionHistory?: AnalyzeInputSessionHistoryItem[];
}

export interface IdeaReviewInput {
  /** @minLength 1 */
  idea: string;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  teamSize?: string | null;
}

export type VenusResponseCardsItemType = typeof VenusResponseCardsItemType[keyof typeof VenusResponseCardsItemType];


export const VenusResponseCardsItemType = {
  analysis: 'analysis',
  market: 'market',
  risk: 'risk',
  roadmap: 'roadmap',
  decision: 'decision',
} as const;

export type VenusResponseCardsItemContent = { [key: string]: unknown };

export type VenusResponseCardsItem = {
  type: VenusResponseCardsItemType;
  title: string;
  content: VenusResponseCardsItemContent;
};

export interface VenusResponse {
  summary: string;
  cards: VenusResponseCardsItem[];
}

export interface GroqKeyStatus {
  configured: boolean;
  /** @nullable */
  maskedKey?: string | null;
}

export interface GroqKeyInput {
  /** @minLength 10 */
  apiKey: string;
}

export interface OnboardingData {
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  teamSize?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  primaryGoal?: string | null;
  completed?: boolean;
}

export interface OnboardingInput {
  companyName: string;
  stage: string;
  industry: string;
  teamSize: string;
  country: string;
  primaryGoal: string;
}

export type ListEventsParams = {
category?: ListEventsCategory;
sentiment?: ListEventsSentiment;
};

export type ListEventsCategory = typeof ListEventsCategory[keyof typeof ListEventsCategory];


export const ListEventsCategory = {
  all: 'all',
  technology: 'technology',
  finance: 'finance',
  markets: 'markets',
  health: 'health',
} as const;

export type ListEventsSentiment = typeof ListEventsSentiment[keyof typeof ListEventsSentiment];


export const ListEventsSentiment = {
  all: 'all',
  positive: 'positive',
  negative: 'negative',
  mixed: 'mixed',
} as const;

export type ListReportsParams = {
category?: ListReportsCategory;
};

export type ListReportsCategory = typeof ListReportsCategory[keyof typeof ListReportsCategory];


export const ListReportsCategory = {
  all: 'all',
  technology: 'technology',
  finance: 'finance',
  markets: 'markets',
  health: 'health',
} as const;

export type ListCompaniesParams = {
category?: ListCompaniesCategory;
sort?: ListCompaniesSort;
};

export type ListCompaniesCategory = typeof ListCompaniesCategory[keyof typeof ListCompaniesCategory];


export const ListCompaniesCategory = {
  all: 'all',
  technology: 'technology',
  finance: 'finance',
  markets: 'markets',
  health: 'health',
} as const;

export type ListCompaniesSort = typeof ListCompaniesSort[keyof typeof ListCompaniesSort];


export const ListCompaniesSort = {
  date: 'date',
  name: 'name',
  impact: 'impact',
} as const;

export type ListThoughtsParams = {
category?: ListThoughtsCategory;
};

export type ListThoughtsCategory = typeof ListThoughtsCategory[keyof typeof ListThoughtsCategory];


export const ListThoughtsCategory = {
  all: 'all',
  technology: 'technology',
  finance: 'finance',
  markets: 'markets',
  health: 'health',
} as const;