# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vera Nexus BI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List causal events chronologically
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [all, technology, finance, markets, health]
        - name: sentiment
          in: query
          schema:
            type: string
            enum: [all, positive, negative, mixed]
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/ripple:
    post:
      operationId: rippleAnalysis
      tags: [events]
      summary: AI ripple analysis for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ripple analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RippleResult"

  # ── REPORTS (Sight page) ──
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List intelligence reports
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [all, technology, finance, markets, health]
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a single intelligence report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}/summary:
    post:
      operationId: reportSummary
      tags: [reports]
      summary: AI summary for a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResult"

  # ── COMPANIES (Crypt page) ──
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies in the corporate graveyard
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [all, technology, finance, markets, health]
        - name: sort
          in: query
          schema:
            type: string
            enum: [date, name, impact]
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a single company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}/autopsy:
    post:
      operationId: companyAutopsy
      tags: [companies]
      summary: AI autopsy analysis for a failed company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Autopsy analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutopsyResult"

  # ── SIGNALS ──
...
  /ai/analyze:
    post:
      operationId: venusAnalyze
      tags: [ai]
      summary: Venus AI business intelligence analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeInput"
      responses:
        "200":
          description: Venus AI JSON card response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenusResponse"
  /ai/idea-review:
    post:
      operationId: ideaReview
      tags: [ai]
      summary: Idea reviewer - evaluate a business idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaReviewInput"
      responses:
        "200":
          description: Idea review result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenusResponse"

  # ── SETTINGS ──
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, year, title, category, sentiment, impact, source]
      properties:
        id:
          type: integer
        year:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [technology, finance, markets, health]
        sentiment:
          type: string
          enum: [positive, negative, mixed]
        impact:
          type: integer
          description: Impact score 0-100
        source:
          type: string
        rippleCount:
          type: integer
        tags:
          type: array
          items:
            type: string
    RippleResult:
      type: object
      required: [eventId, analysis]
      properties:
        eventId:
          type: integer
        analysis:
          type: string
        causalChain:
          type: array
          items:
            type: string
        affectedSectors:
          type: array
          items:
            type: string

    # ── REPORT ──
    Report:
      type: object
      required: [id, title, source, publishedAt, category, summary]
      properties:
        id:
          type: integer
        title:
          type: string
        source:
          type: string
        publishedAt:
          type: string
        category:
          type: string
          enum: [technology, finance, markets, health]
        summary:
          type: string
        imageUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
    SummaryResult:
      type: object
      required: [reportId, summary, keyTakeaways]
      properties:
        reportId:
          type: integer
        summary:
          type: string
        keyTakeaways:
          type: array
          items:
            type: string
        causalImplications:
          type: string

    # ── COMPANY ──
    Company:
      type: object
      required: [id, name, yearRange, description, category]
      properties:
        id:
          type: integer
        name:
          type: string
        yearRange:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [technology, finance, markets, health]
        tags:
          type: array
          items:
            type: string
        failureReason:
          type: ["string", "null"]
    AutopsyResult:
      type: object
      required: [companyId, rootCause, timeline, lessonsLearned]
      properties:
        companyId:
          type: integer
        rootCause:
          type: string
        timeline:
          type: string
        lessonsLearned:
          type: array
          items:
            type: string
        causalChain:
          type: array
          items:
            type: string
        analogy:
          type: ["string", "null"]

    # ── SIGNAL ──
...
    AnalyzeInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        businessContext:
          type: ["string", "null"]
        sessionHistory:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
    IdeaReviewInput:
      type: object
      required: [idea]
      properties:
        idea:
          type: string
          minLength: 1
        stage:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        teamSize:
          type: ["string", "null"]
    VenusResponse:
      type: object
      required: [summary, cards]
      properties:
        summary:
          type: string
        cards:
          type: array
          items:
            type: object
            required: [type, title, content]
            properties:
              type:
                type: string
                enum: [analysis, market, risk, roadmap, decision]
              title:
                type: string
              content:
                type: object

    # ── SETTINGS ──
...