# lib/api-zod/src/generated/types/ideaReviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vera Nexus BI Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface IdeaReviewInput {
  /** @minLength 1 */
  idea: string;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  teamSize?: string | null;
}